Verify signature of the inventory image before downloading plugin bundle
