+++ pkg/discovery/oci_dbbacked.go
@@ -4,12 +4,10 @@
 package discovery
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strconv"
-	"strings"
 
 	"github.com/pkg/errors"
 
@@ -17,9 +15,8 @@ import (
 	"github.com/vmware-tanzu/tanzu-cli/pkg/carvelhelpers"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/common"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/constants"
-	"github.com/vmware-tanzu/tanzu-cli/pkg/cosignhelper"
+	"github.com/vmware-tanzu/tanzu-cli/pkg/cosignhelper/sigverifier"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/plugininventory"
-	"github.com/vmware-tanzu/tanzu-cli/pkg/registry"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/utils"
 	"github.com/vmware-tanzu/tanzu-plugin-runtime/log"
 )
@@ -160,17 +157,10 @@ func (od *DBBackedOCIDiscovery) fetchInventoryImage() error {
 	// The DB has changed and needs to be updated in the cache.
 	log.Infof("Reading plugin inventory for %q, this will take a few seconds.", od.image)
 
-	cosignVerifier, err := od.getCosignVerifier()
+	// Verify the inventory image signature before downloading the plugin inventory database
+	err := sigverifier.VerifyInventoryImageSignature(od.image)
 	if err != nil {
-		return errors.Wrapf(err, "failed to initialize the cosign verifier")
-	}
-
-	if sigVerifyErr := od.verifyInventoryImageSignature(cosignVerifier); sigVerifyErr != nil {
-		log.Warningf("Unable to verify the plugins discovery image signature: %v", sigVerifyErr)
-		// TODO(pkalle): Update the message to convey user to check if they could use the latest public key after we get details of the well known location of the public key
-		errMsg := fmt.Sprintf("Fatal, plugins discovery image signature verification failed. The `tanzu` CLI can not ensure the integrity of the plugins to be installed. To ignore this validation please append %q to the comma-separated list in the environment variable %q.  This is NOT RECOMMENDED and could put your environment at risk!",
-			od.image, constants.PluginDiscoveryImageSignatureVerificationSkipList)
-		log.Fatal(nil, errMsg)
+		return err
 	}
 
 	// download plugin inventory image to get the 'plugin_inventory.db'
@@ -297,63 +287,4 @@ func (od *DBBackedOCIDiscovery) checkDigestFileExistence(hashHexVal, digestPrefi
 		os.Remove(matches[0])
 	}
 	return correctHashFile
-}
-
-func (od *DBBackedOCIDiscovery) verifyInventoryImageSignature(verifier cosignhelper.Cosignhelper) error {
-	signatureVerificationSkipSet := getPluginDiscoveryImagesSkippedForSignatureVerification()
-	if _, exists := signatureVerificationSkipSet[strings.TrimSpace(od.image)]; exists {
-		// log warning message iff user had not chosen to skip warning message for signature verification
-		if skip, _ := strconv.ParseBool(os.Getenv(constants.SuppressSkipSignatureVerificationWarning)); !skip {
-			log.Warningf("Skipping the plugins discovery image signature verification for %q\n ", od.image)
-		}
-		return nil
-	}
-
-	err := verifier.Verify(context.Background(), []string{od.image})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (od *DBBackedOCIDiscovery) getCosignVerifier() (cosignhelper.Cosignhelper, error) {
-	// Get the custom public key path and prepare cosign verifier, if empty, cosign verifier would use embedded public key for verification
-	customPublicKeyPath := os.Getenv(constants.PublicKeyPathForPluginDiscoveryImageSignature)
-
-	registryOptions, err := od.getCosignVerifierRegistryOptions()
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to prepare the registry options for cosign verification")
-	}
-	return cosignhelper.NewCosignVerifier(customPublicKeyPath, registryOptions), nil
-}
-
-// getCosignVerifierRegistryOptions prepares the registry options by including the custom certificate configuration if any
-func (od *DBBackedOCIDiscovery) getCosignVerifierRegistryOptions() (*cosignhelper.RegistryOptions, error) {
-	registryOpts := &cosignhelper.RegistryOptions{}
-	registryName, err := registry.GetRegistryName(strings.TrimSpace(od.image))
-	if err != nil {
-		return nil, err
-	}
-	// get the certificate configuration and update the registry options
-	regCertOptions, err := registry.GetRegistryCertOptions(registryName)
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to get the registry certificate configuration")
-	}
-	registryOpts.CACertPaths = regCertOptions.CACertPaths
-	registryOpts.SkipCertVerify = regCertOptions.SkipCertVerify
-	registryOpts.AllowInsecure = regCertOptions.Insecure
-
-	return registryOpts, nil
-}
-
-func getPluginDiscoveryImagesSkippedForSignatureVerification() map[string]struct{} {
-	discoveryImages := map[string]struct{}{}
-	discoveryImagesList := strings.Split(os.Getenv(constants.PluginDiscoveryImageSignatureVerificationSkipList), ",")
-	for _, image := range discoveryImagesList {
-		image = strings.TrimSpace(image)
-		if image != "" {
-			discoveryImages[image] = struct{}{}
-		}
-	}
-	return discoveryImages
 }
